+++ platformer_tutorial_2025/scripts/scene.py
@@ -1,5 +1,13 @@
+import math
+import os
 import pygame
+import random
+from scripts.clouds import Cloud, Clouds
 from scripts.entities import Enemy
+from scripts.particle import Particle
+from scripts.spark import Spark
+from scripts.tilemap import Tilemap
+from scripts.transitioner import Transitioner
 import sys
 
 class Scene:
@@ -24,28 +32,47 @@ class GameplayScene(Scene):
     '''
     def __init__(self, game, level):
         super().__init__(game)
+        
+        # Metadata
         self.level = level
+        self.complete = False
+
+        # Map stuff
+        self.clouds = Clouds(self.game.assets['clouds'], count=16)
+        self.tilemap = Tilemap(self.game, tile_size=16)
+
+        # Level stuff
+        self.movement = [False, False] # Used to track movement triggers by the player
 
     def load_level(self, map_id):
         '''
         Reset the game on the given level (map_id)
         '''
         # Load map
-        self.game.tilemap.load('data/maps/' + str(map_id) + '.json')
+        self.tilemap.load('data/maps/' + str(map_id) + '.json')
 
         # Handle leaf spawners
         self.leaf_spawners = []
-        for tree in self.game.tilemap.extract([('large_decor', 2)], keep=True):
+        for tree in self.tilemap.extract([('large_decor', 2)], keep=True):
             self.leaf_spawners.append(pygame.Rect(4 + tree['pos'][0], 4 + tree['pos'][1], 23, 13)) # Offset by 4 pixels for leaf falling. Numbers based on tree img size
 
         # Handle original Player/Enemy Spawners
         self.enemies = []
-        for spawner in self.game.tilemap.extract([('spawners', 0), ('spawners', 1)]):
+        for spawner in self.tilemap.extract([('spawners', 0), ('spawners', 1)]):
             if spawner['variant'] == 0: # Player variant
-                self.game.player.pos = spawner['pos']
+                # Spawn player
+                self.game.player.pos = list(spawner['pos'])
                 self.game.player.air_time = 0 # Reset on respawn
+                
+                # Load transition location
+                self.transition_loc = pygame.Rect(spawner['pos'][0], spawner['pos'][1], 8, 15)
             else:
-                self.enemies.append(Enemy(self.game, spawner['pos'], (8, 15), 20, 10))
+                self.enemies.append(Enemy(self.game, spawner['pos'], (8, 15), 10, 10))
+        
+        # Obtain Transitioners
+        self.transitioners = []
+        for tile in self.tilemap.extract([('transitioner', 0)]):
+            self.transitioners.append(Transitioner(self.game, tile['pos'], (8, 15), self.level + 1))
         
         # Reset other entity collections
         self.projectiles = []
@@ -53,16 +80,117 @@ class GameplayScene(Scene):
         self.sparks = []
 
         # Reset values
+        self.game.player.health = self.game.player.max_health
         self.dead = 0 # Number boolean for if player is dead
+        self.complete = False # Are all enemies dead?
+        self.transitioning = False # Did player reach a Transition point?
         self.transition = -30 # Transition speed when moving to a new level
         self.game.scroll = [0, 0] # Offset which emulates a "camera" experience
 
+    def update(self):
+        # Track load level animation
+        if self.transition < 0: # For the start of a level
+            self.transition += 1
+        if self.complete and self.transitioning: # Track contact with transitioner
+            self.transition += 1
+        if self.transition > 30: # Trigger the new level load
+                self.level = min(self.level + 1, len(os.listdir('data/maps')) - 1)
+                self.load_level(self.level)
+
+        # Background assets
+        self.clouds.update()
+
+        for rect in self.leaf_spawners:
+            if random.random() * 49999 < rect.width * rect.height: # Control spawn rate in relation to the size of the tree
+                pos = (rect.x + random.random() * rect.width, rect.y + random.random() * rect.height)
+                self.particles.append(Particle(self.game, 'leaf', pos, velocity=[-0.1, 0.3], frame=random.randint(0, 20)))
+
+        # Check for progression to next level
+        if len(self.enemies) == 0: # All enemies defeated, unlock next room
+            self.complete = True
+            if self.game.player.rect().collidelistall([x.rect() for x in self.transitioners]): # Start the countdown to new level
+                self.transitioning = True
+
+
+        # Handle enemies
+        for enemy in self.enemies.copy():
+            kill = enemy.update(self.tilemap, (0, 0))
+            if kill:
+                self.enemies.remove(enemy)
+
+        # Check Player death
+        if self.dead: # You died, start over in 40 frames
+            self.dead += 1
+            if self.dead >= 10:
+                self.transition = min(30, self.transition + 1)
+            if self.dead > 40:
+                self.load_level(self.level)
+        else: # Update as usual and continue
+            self.game.player.update(self.tilemap, (self.movement[1] - self.movement[0], 0))
+
+        # Resolve sparks
+        for spark in self.sparks.copy():
+            kill = spark.update()
+            if kill:
+                self.sparks.remove(spark)
+        
+        # Resolve Projectiles
+        for projectile in self.projectiles.copy():
+            projectile.move() # TODO: Consider putting move in update
+            projectile.update(self.tilemap)
+
+        # Resolve Particles
+        for particle in self.particles.copy():
+            kill = particle.update()
+            if particle.type == 'leaf':
+                particle.pos[0] += math.sin(particle.animation.frame * 0.035) * 0.3 # Put a wobble on the leaf fall with a sin wave
+            if kill:
+                self.particles.remove(particle)
+
+    def render(self, offset=(0, 0)):
+        # Background
+        self.clouds.render(self.game.display_2, offset=offset)
+
+        # Tilemap
+        self.tilemap.render(self.game.display, offset=offset)
+
+        # Special conditions
+        if self.complete:
+            for tile in self.transitioners:
+                self.game.display.blit(tile.img,
+                    (tile.pos[0] - offset[0],
+                    tile.pos[1] - offset[1])  
+                )
+
+        # Enemies
+        for enemy in self.enemies:
+            enemy.render(self.game.display, offset=offset)
+        
+        # Player
+        if not self.dead:
+            self.game.player.render(self.game.display, offset=offset)
+
+        # Sparks
+        for spark in self.sparks:
+            spark.render(self.game.display, offset=offset)
+
+        # Projectiles
+        for projectile in self.projectiles:
+            projectile.render(self.game.display, offset=offset)
+
+        # Particles
+        for particle in self.particles:
+            particle.render(self.game.display, offset=offset)
     
 
 class PauseScene(Scene):
     def __init__(self, game):
         super().__init__(game)
         self.pause_display = pygame.Surface((self.game.screen.get_width() // 4, self.game.screen.get_height() // 4)) # Set pause space
+        self.is_paused = False
+
+    def update(self):
+        pass # To be used for any updates later
 
     def handle_input(self):
         for event in pygame.event.get(): # event is where the... events get stored
@@ -71,11 +199,13 @@ class PauseScene(Scene):
                     sys.exit()
                 if event.type == pygame.KEYDOWN:
                     if event.key == pygame.K_ESCAPE:
-                        self.game.is_paused = not self.game.is_paused
+                        self.pause()
 
-    def render_pause(self):
+    def render(self):
         # TODO: This is centered correctly, but I need a proper pause screen...
         pygame.draw.circle(self.pause_display, (255, 255, 255), (self.pause_display.get_width() // 2, self.pause_display.get_height() // 2), self.pause_display.get_width() // 3)
         self.game.screen.blit(self.pause_display, (self.game.screen.get_width() // 2 - self.pause_display.get_width() // 2, self.game.screen.get_height() // 2 - self.pause_display.get_height() // 2))
 
+    def pause(self):
+        self.is_paused = not self.is_paused