+++ platformer_tutorial_2025/changelog.md
@@ -1,3 +1,7 @@
+## 12/13/2025
+- Refactored game to have Game and Scene objects
+- Added Transition Point instead of auto-level transition
+
 ## 12/3/2025
 - Added health and damage system
 - Cleaned repo and prepared plan for a "complete" game